+++ packages/extension-koni-ui/src/Popup/Home/Tokens/UpperBlock.tsx
@@ -1,10 +1,12 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { balanceNoPrefixFormater } from '@subwallet/extension-base/utils';
 import { useSelector, useTranslation } from '@subwallet/extension-koni-ui/hooks';
 import { reloadCron, saveShowBalance } from '@subwallet/extension-koni-ui/messaging';
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
-import { Button, Icon, Number, SwNumberProps, Tag } from '@subwallet/react-ui';
+import { formatBalance } from '@subwallet/extension-koni-ui/utils';
+import { Button, formatNumber, Icon, Number, SwNumberProps, Tag, Tooltip } from '@subwallet/react-ui';
 import CN from 'classnames';
 import { ArrowsClockwise, ArrowsLeftRight, CopySimple, Eye, EyeSlash, PaperPlaneTilt, ShoppingCartSimple } from 'phosphor-react';
 import React, { useCallback, useState } from 'react';
@@ -36,6 +38,7 @@ function Component (
   const { t } = useTranslation();
   const { isShowBalance } = useSelector((state) => state.settings);
   const [reloading, setReloading] = useState(false);
+  const { currencyData } = useSelector((state) => state.price);
 
   const onChangeShowBalance = useCallback(() => {
     saveShowBalance(!isShowBalance).catch(console.error);
@@ -53,21 +56,34 @@ function Component (
   return (
     <div className={`tokens-upper-block ${className} ${isShrink ? '-shrink' : ''}`}>
       <div className='__total-balance-value-container'>
-        <div
-          className='__total-balance-value-content'
-          onClick={isShrink ? onChangeShowBalance : undefined}
+        <Tooltip
+          overlayClassName={CN('__currency-value-detail-tooltip', {
+            'ant-tooltip-hidden': !isShowBalance
+          })}
+          placement='top'
+          title={formatNumber(totalValue, 0, balanceNoPrefixFormater) + ' ' + currencyData.symbol}
         >
-          <Number
-            className={'__total-balance-value'}
-            decimal={0}
-            decimalOpacity={0.45}
-            hide={!isShowBalance}
-            prefix='$'
-            size={38}
-            subFloatNumber
-            value={totalValue}
-          />
-        </div>
+          <div
+            className='__total-balance-value-content'
+            onClick={isShrink ? onChangeShowBalance : undefined}
+          >
+            <Number
+              className={'__total-balance-value'}
+              decimal={0}
+              decimalOpacity={0.45}
+              hide={!isShowBalance}
+              size={38}
+              subFloatNumber
+              value={totalValue}
+            />
+            {isShowBalance && <div className={CN('__total-balance-symbol', {
+              '-not-show-balance': isShrink && formatBalance(totalValue, 0).length > 10
+            })}
+            >
+              {currencyData.symbol}
+            </div>}
+          </div>
+        </Tooltip>
       </div>
       {!isShrink && (
         <div className={'__balance-change-container'}>
@@ -75,7 +91,7 @@ function Component (
             className='button-change-show-balance'
             icon={(
               <Icon
-                phosphorIcon={ !isShowBalance ? Eye : EyeSlash}
+                phosphorIcon={!isShowBalance ? Eye : EyeSlash}
               />
             )}
             onClick={onChangeShowBalance}
@@ -88,7 +104,8 @@ function Component (
             decimal={0}
             decimalOpacity={1}
             hide={!isShowBalance}
-            prefix={isPriceDecrease ? '- $' : '+ $'}
+            prefix={isPriceDecrease ? `- ${(currencyData.isPrefix && currencyData.symbol) || ''}` : `+ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+            suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
             value={totalChangeValue}
           />
           <Tag
@@ -256,6 +273,25 @@ export const UpperBlock = styled(Component)<Props>(({ theme: { token } }: Props)
       width: token.size
     },
 
+    '.__total-balance-value-content': {
+      display: 'flex',
+      justifyContent: 'center',
+      cursor: 'pointer',
+      width: 'fit-content',
+      margin: 'auto'
+    },
+
+    '.__total-balance-symbol': {
+      marginLeft: -2,
+      fontSize: token.fontSizeSM,
+      lineHeight: token.lineHeightHeading6,
+
+      '&.-not-show-balance': {
+        display: 'none'
+      }
+
+    },
+
     '&.-shrink': {
       paddingBottom: 32,
       flexDirection: 'row',
@@ -266,7 +302,8 @@ export const UpperBlock = styled(Component)<Props>(({ theme: { token } }: Props)
 
       '.__total-balance-value-content': {
         cursor: 'pointer',
-        width: 'fit-content'
+        width: 'fit-content',
+        margin: 0
       },
 
       '.__total-balance-value': {
@@ -291,5 +328,6 @@ export const UpperBlock = styled(Component)<Props>(({ theme: { token } }: Props)
         width: token.sizeXS
       }
     }
+
   });
 });