+++ packages/extension-koni-ui/src/components/Modal/Swap/ChooseFeeTokenModal.tsx
@@ -3,11 +3,13 @@
 
 import { swapCustomFormatter } from '@subwallet/extension-base/utils';
 import ChooseFeeItem from '@subwallet/extension-koni-ui/components/Field/Swap/ChooseFeeItem';
+import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { ModalContext, Number, SwModal } from '@subwallet/react-ui';
 import BigN from 'bignumber.js';
 import CN from 'classnames';
 import React, { useCallback, useContext } from 'react';
+import { useSelector } from 'react-redux';
 import styled from 'styled-components';
 
 type Props = ThemeProps & {
@@ -21,7 +23,7 @@ const numberMetadata = { maxNumberFormat: 8 };
 
 const Component: React.FC<Props> = (props: Props) => {
   const { className, estimatedFee, items, modalId, onSelectItem, selectedItem } = props;
-
+  const { currencyData } = useSelector((state: RootState) => state.price);
   const { inactiveModal } = useContext(ModalContext);
 
   const onCancel = useCallback(() => {
@@ -48,8 +50,9 @@ const Component: React.FC<Props> = (props: Props) => {
               decimalOpacity={0.45}
               formatType={'custom'}
               metadata={numberMetadata}
-              prefix={'$'}
+              prefix={(currencyData.isPrefix && currencyData.symbol) || ''}
               size={30}
+              suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
               value={estimatedFee}
             />
             <span className={'__pay-with'}>Pay with token:</span>